Factor out Case and Cocase
