+++ lang/fun/src/syntax/terms/mod.rs
@@ -1,13 +1,9 @@
-use std::rc::Rc;
-
 use codespan::Span;
 use derivative::Derivative;
-use printer::{
-    theme::ThemeExt,
-    tokens::{CASE, COCASE, DOT, FAT_ARROW},
-    DocAllocator, Print,
-};
+use printer::{tokens::FAT_ARROW, DocAllocator, Print};
 
+mod case;
+mod cocase;
 mod constructor;
 mod destructor;
 mod fun;
@@ -19,6 +15,8 @@ mod local_let;
 mod op;
 mod paren;
 mod var;
+pub use case::*;
+pub use cocase::*;
 pub use constructor::*;
 pub use destructor::*;
 pub use fun::*;
@@ -31,7 +29,7 @@ pub use op::*;
 pub use paren::*;
 pub use var::*;
 
-use super::{context::TypingContext, print_cases, Name};
+use super::context::TypingContext;
 
 // Clause
 //
@@ -69,214 +67,6 @@ impl<T: Print> Print for Clause<T> {
                 .append(alloc.space())
                 .append(self.rhs.print(cfg, alloc))
         }
-    }
-}
-
-// Case
-//
-//
-
-#[derive(Derivative, Debug, Clone)]
-#[derivative(PartialEq, Eq)]
-pub struct Case {
-    #[derivative(PartialEq = "ignore")]
-    pub span: Span,
-    pub destructee: Rc<Term>,
-    pub cases: Vec<Clause<Name>>,
-}
-
-impl Print for Case {
-    fn print<'a>(
-        &'a self,
-        cfg: &printer::PrintCfg,
-        alloc: &'a printer::Alloc<'a>,
-    ) -> printer::Builder<'a> {
-        self.destructee
-            .print(cfg, alloc)
-            .append(DOT)
-            .append(alloc.keyword(CASE))
-            .append(alloc.space())
-            .append(print_cases(&self.cases, cfg, alloc))
-    }
-}
-
-impl From<Case> for Term {
-    fn from(value: Case) -> Self {
-        Term::Case(value)
-    }
-}
-
-#[cfg(test)]
-mod case_tests {
-    use codespan::Span;
-    use printer::Print;
-
-    use crate::{
-        parser::fun,
-        syntax::{context::ContextBinding, types::Ty},
-    };
-
-    use super::{Case, Clause, Lit, Term, Var};
-    use std::rc::Rc;
-
-    fn example_empty() -> Case {
-        Case {
-            span: Span::default(),
-            destructee: Rc::new(Var::mk("x").into()),
-            cases: vec![],
-        }
-    }
-
-    fn example_tup() -> Case {
-        Case {
-            span: Span::default(),
-            destructee: Rc::new(Var::mk("x").into()),
-            cases: vec![Clause {
-                span: Span::default(),
-                xtor: "Tup".to_owned(),
-                context: vec![
-                    ContextBinding::TypedVar {
-                        var: "x".to_string(),
-                        ty: Ty::mk_int(),
-                    },
-                    ContextBinding::TypedVar {
-                        var: "y".to_string(),
-                        ty: Ty::mk_int(),
-                    },
-                ],
-                rhs: Term::Lit(Lit::mk(2)),
-            }],
-        }
-    }
-
-    #[test]
-    fn display_empty() {
-        assert_eq!(
-            example_empty().print_to_string(Default::default()),
-            "x.case { }"
-        )
-    }
-
-    #[test]
-    fn parse_empty() {
-        let parser = fun::TermParser::new();
-        assert_eq!(parser.parse("x.case { }"), Ok(example_empty().into()));
-    }
-
-    #[test]
-    fn display_tup() {
-        assert_eq!(
-            example_tup().print_to_string(Default::default()),
-            "x.case { Tup(x : Int, y : Int) => 2 }"
-        )
-    }
-
-    #[test]
-    fn parse_tup() {
-        let parser = fun::TermParser::new();
-        assert_eq!(
-            parser.parse("x.case { Tup(x : Int, y : Int) => 2 }"),
-            Ok(example_tup().into())
-        );
-    }
-}
-
-// Cocase
-//
-//
-
-#[derive(Derivative, Debug, Clone)]
-#[derivative(PartialEq, Eq)]
-pub struct Cocase {
-    #[derivative(PartialEq = "ignore")]
-    pub span: Span,
-    pub cocases: Vec<Clause<Name>>,
-}
-
-impl Print for Cocase {
-    fn print<'a>(
-        &'a self,
-        cfg: &printer::PrintCfg,
-        alloc: &'a printer::Alloc<'a>,
-    ) -> printer::Builder<'a> {
-        alloc
-            .keyword(COCASE)
-            .append(alloc.space())
-            .append(print_cases(&self.cocases, cfg, alloc))
-    }
-}
-
-impl From<Cocase> for Term {
-    fn from(value: Cocase) -> Self {
-        Term::Cocase(value)
-    }
-}
-
-#[cfg(test)]
-mod cocase_tests {
-    use codespan::Span;
-    use printer::Print;
-
-    use crate::parser::fun;
-
-    use super::{Clause, Cocase, Lit, Term};
-
-    fn example_empty() -> Cocase {
-        Cocase {
-            span: Span::default(),
-            cocases: vec![],
-        }
-    }
-
-    fn example_stream() -> Cocase {
-        Cocase {
-            span: Span::default(),
-            cocases: vec![
-                Clause {
-                    span: Span::default(),
-                    xtor: "Hd".to_owned(),
-                    context: vec![],
-                    rhs: Term::Lit(Lit::mk(2)),
-                },
-                Clause {
-                    span: Span::default(),
-                    xtor: "Tl".to_owned(),
-                    context: vec![],
-                    rhs: Term::Lit(Lit::mk(4)),
-                },
-            ],
-        }
-    }
-
-    #[test]
-    fn display_empty() {
-        assert_eq!(
-            example_empty().print_to_string(Default::default()),
-            "cocase { }"
-        )
-    }
-
-    #[test]
-    fn parse_empty() {
-        let parser = fun::TermParser::new();
-        assert_eq!(parser.parse("cocase { }"), Ok(example_empty().into()));
-    }
-
-    #[test]
-    fn display_stream() {
-        assert_eq!(
-            example_stream().print_to_string(Default::default()),
-            "cocase {\n    Hd => 2,\n    Tl => 4\n}"
-        )
-    }
-
-    #[test]
-    fn parse_stream() {
-        let parser = fun::TermParser::new();
-        assert_eq!(
-            parser.parse("cocase { Hd => 2, Tl => 4 }"),
-            Ok(example_stream().into())
-        );
     }
 }
